Move `__ensure_current_device` to own header
